requirements.txt
flask==2.3.3
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2.2
shap==0.42.1
lime==0.2.0.1
plotly==5.17.0
dash==2.14.1
dash-bootstrap-components==1.5.0
joblib==1.3.2
seaborn==0.12.2
matplotlib==3.7.2
werkzeug==2.3.7
python-dotenv==1.0.0
nltk==3.8.1
opencv-python==4.8.1.78
Pillow==10.0.1
scikit-image==0.21.0
requests==2.31.0
